pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "datapalette"
version = "0.1.0"
description = "Image dataset preprocessing and augmentation toolkit with sklearn-compatible transforms."
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    {name = "Juan Garassino"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Image Processing",
]
dependencies = [
    "numpy",
    "opencv-python",
    "tqdm",
    "pyyaml",
    "scikit-learn",
]

[project.optional-dependencies]
torch = ["torch", "torchvision"]
dev = ["pytest", "pytest-cov", "ruff", "mypy"]

[project.scripts]
datapalette = "datapalette.cli:main"

[tool.setuptools.packages.find]
include = ["datapalette*"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.9"
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
